requirements.txt
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
cachetools==5.3.2
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.14.1
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
